Profile Widget added to global ToolSettings struct and edit mode bevel tool.
Now the same profile widget should be accessible everywhere, so the changes
should carry over from one call to the next.

The tool settings for bevel are now drawn in a custom callback which also enables
some fancier hiding of properties when they're not being used, which I've also
taken advantage of in this commit.

There are still a couple bugs with the new redo panel but the bulk of the work is
finished for making the bevel tool work with custom profiles.