GPencil: Fix problem when color is nofill in SVG
If the color in the SVG converted was nofill, the stroke had fill color in grease pencil and this creates glitches because the shape was not defined to be filled.

Now, the name of the color is checked to verify if the material is for stroke, and if only there is one material with this name, the fill is disabled.