USD: Fixed issue with mesh particle export
When mesh particles were exported, the object name was still used for each
instance of that object. As such, only one instance was written to USD.
This is now fixed by suffixing the object name with the particle's
persistent ID, giving each particle XForm a unique name.

Particles are still only written when they are alive, which means that they
are always visible (there is currently no code that deals with marking them
as invisible outside their lifespan).