Outliner: Show active improvements
Prevent the show active operator from scrolling beneath the
allowed scrolling range. This removes jumpy redraws after the
operator executes.

The horizontal centering is difficult, because the elements
outside the view of the outliner do not have the xend properly set
and the max width of the outliner is not properly calculated.
This means it is not possible to center the x position while
keeping the view in range of scrolling.

This could be resolved by drawing the outliner after setting the y
position, but that is not an ideal solution.