Outliner: Allow clicking on row icons
The feature to click on row icons was removed (presumably because
the icons now draw stacked with a number, rather than individually).
This brings back support for clicking because the benefits of
clicking on the icon without opening the element outweighs the con
of not being able to select grouped elements. A solution for that
is being worked on.