USD: Experimental support for instancing/referencing
Dupli-object meshes are now written to USD as references to the origina
mesh. This is still very limited in correctness, as there are issues
referencing to materials from a referenced mesh.

I am still committing this, as it also introduces some code clarifications
and gives us a place to start when continuing the quest for proper
instancing in USD.