GPencil: Enable gradient factor for Line strokes
Now it's possible to use the same gradient used in Dots and Boxes with Line strokes. This was disabled, pending to solve the overlap glitches, but as this was fixed with stencil, we can enable it again.

Also, the gradient formula has been changed to use the same of Dots.