Move filter settings into popup
Saves space, hides complexity and is more consistent to other editors
(e.g. viewport and outliner). The search button is in the popup too for
now, we may want to move that later.
Note the filter popup is not located in the header, but in the directory
path (`UI`) region.

Had to do two additional changes:
* Allow registering a panel without drawing it in the default location.
  Required so we can register the popup panel in the `UI` region without
  it drawing there. For other places that's no issue because they are
  registered for headers - which don't draw panels anyway.
* `UI` regions are used for category tabs by default. Added a flag to
  allow region types to disable categories entirely.