Draw column header in short-list view
Similar to what you know from other file browsers, at the table top
there's a row indicating column names and showing a little triangle to
mark which column is used for sorting. Later, this will become
interactive (to allow selecting sort method by clicking on a column
name).

I decided to hardcode drawing of this (like the rest of the file
browser) and just offset the file layout a bit. However that requires
some hacks for View2D and that the offset is respected in multiple
places. So now I think using a separate region instead makes more sense.
Leaving this as is now, code quality for file browser is pretty low
anyway.