Outliner: Support click+drag for toggling disclosure triangles
This reworks the openclose operator for toggling disclosure
triangles. Rather than mapping to `Enter` by default, it supports
left click and left tweak events. This allows click+drag over
elements in the tree to open or close levels. Once the click+drag
is started, only elements on that level in the outliner will open
or close.

This removes the toggling from `outliner_select.c` which adds more
clarity to the activate function.