USD: Allow user to disable export of UV maps and mesh normals
The UV maps and mesh normals take up a significant amount of space, so
it's better to make them optional. They're still enabled by default,
though.

Comparison: a shot of Spring (03_035_A to be exact) is 1.2 GiB  when
exported with UVs and normals, and 262 MiB without. We probably have
room for optimisation of written UVs and normals.