Outliner: Improve click and drag for openclose
This adds the ability to click and drag over a disclosure triangle
and immediately drag back over it to close it. Previously it
required opening the next element before closing the first opened
element because of the way the operator stored the previously
toggled disclosure. This makes the operator more easy to use and
understand.