Outliner: Fix small issues in walk navigation
After adding open/close features to walk navigation, a few small
issues appeared. This adds a few more checks to ensure a walk
or expand really occurs when the desired keys are pressed, rather
than expanding when a walk is desired.

Also moved the walking code to separate functions to make the
switch statement more clear to read.