Bevel cutoff vertex mesh method: New profile height method.
The height of the transformed profile is calculated when cutoff profiles
are enabled and then used later for the length of the sides of the cutoff
faces in the vertex mesh. The lengths can still be incorrect for more
acute angles between incoming edges, but the solution to that will probably
be related to the solution for the directions of the cutoff face sides.