Outliner: Refactor show active to show all instances of object
Previously the show active operator opened the tree to show the
first instance of the active object. With collection linking,
the active object could be in multiple collections. This expands
the show active operator to expand all hierarchies that contain
the active object. The view will be centered on the first
active object in the outliner.