Allow unavailable sockets to be taged as "in use".
Unavailable sockets can and may still be linked, not tagging them as "in use" implies that they aren't linked, which is not true. This causes problems when the developer use the SOCK_IN_USE flag as an indication of the linking status of a socket. An example for this would be versioning code.