Reconfigure the outputs of the Noise node.
We changed the order of outputs so that the Value output is first. We did that because users mostly use the first output, and since noise is most commonly used as a bump texture, roughness textures, or to control other non-color data, it make sense that the first output should be the Value one. Previously, most users used the Color output for practically everything just because it is first in the output list. So reordering should save users some render time without them realizing.