Outliner: walk navigation operator
Adds a keyboard walk navigation and selection operator to the
outliner. Up and down arrow keys walk up and down the list of
elements, and left and right walk in and out or open and close
elements if the elements are open or closed.

Also moves opencolose toggling to the openclose operator to
remove duplicate code.

Also allows click+drag for openclose of element subtrees.

The outliner tree building is tweaked slightly to set the proper
parents in scene display mode for walk select to function
correctly.