GPencil: Create Linear gradient fill colors in SVG conversion
There are several limitations in this conversion because SVG support multiple color, but GP only 2. For GP only the extreme colors (first and last) are used.

The mix factor is always set to 0.5.

The angle of the gradient is calculated using the direction of the line defined in Linear gradient.