EditMesh support mirror on all axes
Part of T68930

This patch adds two more mirror options that can be enabled simultaneously: Mirror Y and Z.

Since different transformation options use their own mirroring code (eg edge / vert slide and proportional editing), this patch proposes a generic solution that can be applied to any transformation.

The idea is to create a new array type that points to the location of each reference element with the mirror element location.

This reduces the amount of `TransData` that is already overkill.

Reviewers: campbellbarton

Differential Revision: https://developer.blender.org/D5720