GPencil: Redesign input samples management
Now, instead to generate fake events in a linear interpolation and try to smooth the line later, an arc is created between the points and try to reproduce the ballistic trajectory of the mouse/pen. This gets better results and especially when the mouse moves very fast getting a more organic result.

Also, the input samples maximum has been changed to 15.

Note: Still we need to test DOT materials to define if the samples are getting good result of we must make more dense the number of points for these materials.