GPencil: New Vertex Paint Replace Tool
This tool replaces the old mixed color with the current one, but keeps the same mix factor. This is not possible with Draw tool, because the draw tool replaces the mix factor.

With this tool is very easy tint previous colors, but keeping the same mix and pressure values to maintain the same artistic look and feel.