Make box selection Include Handles option only act on visible handles
The option would previously act on invisible handles. Because of this,
it wouldn't be wise to enable it by default, even though it can be
tremendously useful (esp. with `Only Selected Keyframes Handles`
disabled).
This changes things based on animator feedback, so that we can enable
the option by default.

I can see that for some people even the old behavior of acting on
invisible handles could be useful, so we might consider bringing that
option back. `Include Handles` could then become an enum to chose the
behavior.