Always apply handle type change to both handles when key is selected.
Previously we only applied the change to both handles if only the key
but no handle is selected.
Based on feedback by Hjalti, this new behavior is more practicable.
Cases where you'd want only one handle to be changed are really rare (he
can't think of one himself) so rather make the common case work well.

This wasn't much of an issue before, as selecting a key always selected
the handles with it, and box select had `Include Handles` disabled by
default.