CMake/GTest: allow defining CLI arguments for GTEST tests
For USD tests I need to load a blend file from a GTest test, and in order to do that properly it needs to get the location of the lib/tests directory. The easiest way to do this is to pass it on the CLI.

To make it possible to extend the `BLENDER_SRC_GTEST_EX()` without adding yet more macros, it now accepts keyword arguments.

Differential Revision: https://developer.blender.org/D6236