Overlay Mode: Fix infront depth sorting
This is an improvement to 2.81. In front objects and extras are correctly
depth sorted.

The approach is to share an in-front depth buffer between all engines
so that workbench fill it with the infront objects depth. Then the overlay
engine can use it again to draw in front objects overlays correctly
occluded by the other "in front" objects surfaces.