USD: Check location of USD JSON files in unit tests
This commit adds a test to see whether the USD JSON files can be found
and loaded. Simply the ability to create a USD Stage for a specific
filename means that the extension has been recognised by the USD
library, and that a USD plugin has been loaded to write such files.

USD has been patched to not only look in `{exe_directory}/lib/usd`, but
also in `{exe_directory}/../lib/usd`; the latter is required to resolve
the path for test executables in `bin/tests/`.