GPencil: Refactor: Add new logic to create batches and store draw passes
We introduce a new set of temp objects (GPENCIL_tVfx, GPENCIL_tLayer,
GPENCIL_tObject) allocated in BLI_memblocks to allow fast traversal and
easy memory management.

We also introduce an iterator to iterate over visible strokes of an object
to avoid code duplication.

Strokes are all in one batch and the fill batch will only contain indices
to the stroke buffer.

For now the objects are only drawn at the center only with white 1px stroke
and all strokes are not interupted (the finish at center).