GPencil: Refactor: Add basic Material color support via UBO
For now there is at least one UBO per object. It is kept relatively small
to avoid overhead. When we implement sharing of UBOs accross objects we
could increase the material count per UBO and reduce the binding overhead.

This also put the BLI_memblocks in the ViewLayer instead of the viewport
storage list. This is because we need to correctly free UBOs when blender
quits. This can become a problem in the future but there are other places
like this in other engines so we should fix them all at once.