USD: patch USD to allow Blender to construct a path to JSON files
USD requires some JSON files, and it uses a static constructor to
determine the possible filesystem paths to find those files. This made
it impossible for Blender to pass a path to the USD library at runtime,
as the constructor would run before Blender's main() function. We have
patched USD (see usd.diff) to avoid that particular static constructor,
and have an initialisation function instead.