Add positional tracking option (3DoF vs. 6DoF)
For passing around the HMD or preparing scenes for 3DoF interaction
(e.g. cardboards) this option is quite handy.

After feedback from @sebastian_k and @noemis, decision was that
disabling the positional tracking should not reset the position to the
starting one. It should just keep whatever is the current position.
So we have to get the local pose (which is different from the per eye
pose) out of OpenXR so we can have more dynamic base position that
changes depending on the positional tracking option.

Not that this isn't working 100% as wanted: There's a slight offset when
disabling the positional tracking. I don't want to spend too much time
on something minor-ish while there are more major features to get done.