GPencil: Refactor: Implement Stroke self overlap option
This is a novel approach were we use the depth buffer instead of the
stencil buffer. This allows up to ~1e24 different indices to draw to.
We use a DRW_STATE_DEPTH_GREATER to discard the fragment comming from the
same stroke.

Since we can have strokes that don't have this option, we use the point
indices instead of the stroke index to avoid changing the state during
drawing.