GPencil: Refactor: Implement Opacity Masking
This is better than previous implementation but has a higher cost.

We render masked layer to a RGBA16F buffer and we then mask the revealage
and the color in a 2 pass blending.

This could maybe optimized a bit to reduce to 1 blending on recent hardware
but the cost will still be way higher than the stencil method.

On the plus side, there is less code change than with the stencil.