GPencil: First step to fix damaged UI by Brush panel refactor done in master
The patch done in master was done before the Sculpt and Weight brushes were real brushes and also before the Vertex Paint mode was implemented. As result of that, the merge of master has broken almost all panels of brushes and all Vertex Paint mode.

This commit only fix part of the damages.