GPencil: Refactor: Add Swirl FX
The angle is no longer mulitplied by 8. We need to change the UI and patch
old files to preserve an appropriate slider feeling.

The Swirl FX (like others) is now affected by distance of the source object.

Transparent option has not been reimplemented as you can use masking for
that.