GPencil: Refactor: Use normal approximation for lighting
We compute the normal using the bounding box size and the view vector to
the center of the bounding box. We then scale the normal w.r.t. the BBox
size and transform it back to world space. This gives us a flat normal
for objects without depth and a rotating normal for object that are
not.

This will also be used for correct depth mixing.