GPencil: Refactor: Implement rounded caps
This was a bit tricky to do it cleanly. We create a mask using the stroke
point coordinates and extend the stroke geometry to uncover the mask if the
round caps are needed.

This was needed to be able to workaround the projection artifact of uvs.

Note that this code could be used for stroke shape in the future.